fix(rename): skip tagged/anchored node_id scalars; move ReDoS perf pins to the perf suite
Review findings on #404: an explicit tag or anchor sits outside the
scalar's source range, so the splice kept it while swapping the value
under it — the old regex never unwrapped those shapes, so restore that
parity with a guard. The two wall-clock ReDoS assertions move from the
parallel unit suite into the dedicated perf config (singleFork, retry: 1)
that exists for exactly this kind of budget.
